// src/package/components/pa-time/type.d.ts
export type MDatePickerType =
  | "date-picker-group"
  | "date-picker"
  | "date-time-picker-group"
  | "date-time-picker"
  | "month-picker-group"
  | "month-picker"
  | "time-picker-group"
  | "time-picker"
  | "year-picker-group"
  | "year-picker";